Add test suite for Importing Process
For https://github.com/instedd/maap-collector/issues/317

This commit adds unit tests for importing process (`LabRecordImport` and `ElectronicPharmacyStockRecord` entities). Head over to the aforementioned issue for more details.

This initial test suite was designed and crafted to be minimal due to budget constraints. It specifically aims to cover the importing process in the backend, which has been the cause of the majority of the bugs so far. I consider this as the first iteration towards a clean and well maintained unit testing suite.

There are still a lot of test cases worthy to be added in places such as controllers and models. It will be great to do that in the near future.